requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.

pbr>=1.6

oslo.config>=2.3.0
pecan>=0.8.0
PasteDeploy>=1.5.0
Werkzeug>=0.7
WSME>=0.7
